Split Destroy Admin Tests
