+++ decidim-admin/lib/decidim/admin/test.rb
@@ -10,3 +10,4 @@ require "decidim/admin/test/filterable_examples"
 require "decidim/admin/test/filters_participatory_space_users_examples"
 require "decidim/admin/test/filters_participatory_space_user_roles_examples"
 require "decidim/admin/test/manage_hide_content_examples"
+require "decidim/admin/test/destroy_admin_examples"
